requirements.txt
flask
flasgger
flask_jwt_extended
psycopg2-binary
flasgger
flask_cors
werkzeug

